new tests for endpoint
